Move mismatch checks to own file
