+++ src/lib/filter/adjudicators/postCollection.ts
@@ -2,41 +2,10 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 
 import { Event, Operation } from "../../enums";
-import { AdmissionRequest, Binding } from "../../types";
-import { __, allPass, any, anyPass, curry, difference, equals, gt, length, not, nthArg, pipe } from "ramda";
+import { __, allPass, anyPass, curry, difference, equals, gt, length, not, nthArg, pipe } from "ramda";
 import { KubernetesObject } from "kubernetes-fluent-client";
-import {
-  definedAnnotations,
-  definedEvent,
-  definedGroup,
-  definedKind,
-  definedLabels,
-  definedName,
-  definedNameRegex,
-  definedNamespaceRegexes,
-  definedNamespaces,
-  definedVersion,
-  definesAnnotations,
-  definesDelete,
-  definesDeletionTimestamp,
-  definesGroup,
-  definesKind,
-  definesLabels,
-  definesName,
-  definesNameRegex,
-  definesNamespaceRegexes,
-  definesNamespaces,
-  definesVersion,
-} from "./binding";
-import {
-  carriedAnnotations,
-  carriedLabels,
-  carriedName,
-  carriedNamespace,
-  carriesNamespace,
-  missingDeletionTimestamp,
-} from "./kubernetesObject";
-import { declaredOperation, declaredGroup, declaredVersion, declaredKind } from "./admissionRequest";
+import { definedKind, definedNamespaces, definesDelete, definesDeletionTimestamp, definesNamespaces } from "./binding";
+import { carriedNamespace, carriesNamespace } from "./kubernetesObject";
 
 /*
   Naming scheme:
@@ -48,78 +17,12 @@ import { declaredOperation, declaredGroup, declaredVersion, declaredKind } from 
 /*
   post-collection comparitors
 */
-export const mismatchedDeletionTimestamp = allPass([
-  pipe(nthArg(0), definesDeletionTimestamp),
-  pipe(nthArg(1), missingDeletionTimestamp),
-]);
-
-export const mismatchedName = allPass([
-  pipe(nthArg(0), definesName),
-  pipe((binding, kubernetesObject) => definedName(binding) !== carriedName(kubernetesObject)),
-]);
-
-export const mismatchedNameRegex = allPass([
-  pipe(nthArg(0), definesNameRegex),
-  pipe((binding, kubernetesObject) => new RegExp(definedNameRegex(binding)).test(carriedName(kubernetesObject)), not),
-]);
 
 export const bindsToKind = curry(
   allPass([pipe(nthArg(0), definedKind, equals(""), not), pipe((binding, kind) => definedKind(binding) === kind)]),
 );
 export const bindsToNamespace = curry(pipe(bindsToKind(__, "Namespace")));
 export const misboundNamespace = allPass([bindsToNamespace, definesNamespaces]);
-
-export const mismatchedNamespace = allPass([
-  pipe(nthArg(0), definesNamespaces),
-  pipe((binding, kubernetesObject) => definedNamespaces(binding).includes(carriedNamespace(kubernetesObject)), not),
-]);
-
-export const mismatchedNamespaceRegex = allPass([
-  pipe(nthArg(0), definesNamespaceRegexes),
-  pipe((binding, kubernetesObject) =>
-    pipe(
-      any((regEx: string) => new RegExp(regEx).test(carriedNamespace(kubernetesObject))),
-      not,
-    )(definedNamespaceRegexes(binding)),
-  ),
-]);
-
-export const metasMismatch = pipe(
-  (defined, carried) => {
-    const result = { defined, carried, unalike: {} };
-
-    result.unalike = Object.entries(result.defined)
-      .map(([key, value]) => {
-        const keyMissing = !Object.hasOwn(result.carried, key);
-        const noValue = !value;
-        const valMissing = !result.carried[key];
-        const valDiffers = result.carried[key] !== result.defined[key];
-
-        // prettier-ignore
-        return (
-          keyMissing ? { [key]: value } :
-          noValue ? {} :
-          valMissing ? { [key]: value } :
-          valDiffers ? { [key]: value } :
-          {}
-        )
-      })
-      .reduce((acc, cur) => ({ ...acc, ...cur }), {});
-
-    return result.unalike;
-  },
-  unalike => Object.keys(unalike).length > 0,
-);
-
-export const mismatchedAnnotations = allPass([
-  pipe(nthArg(0), definesAnnotations),
-  pipe((binding, kubernetesObject) => metasMismatch(definedAnnotations(binding), carriedAnnotations(kubernetesObject))),
-]);
-
-export const mismatchedLabels = allPass([
-  pipe(nthArg(0), definesLabels),
-  pipe((binding, kubernetesObject) => metasMismatch(definedLabels(binding), carriedLabels(kubernetesObject))),
-]);
 
 /*
  * If the object does not have a namespace, and it is not a namespace,
@@ -182,25 +85,4 @@ export const operationMatchesEvent = anyPass([
   pipe(nthArg(1), equals(Event.ANY)),
   pipe((operation: Operation, event: Event): boolean => operation.valueOf() === event.valueOf()),
   pipe((operation: Operation, event: Event): boolean => (operation ? event.includes(operation) : false)),
-]);
-
-export const mismatchedEvent = pipe(
-  (binding: Binding, request: AdmissionRequest): boolean =>
-    operationMatchesEvent(declaredOperation(request), definedEvent(binding)),
-  not,
-);
-
-export const mismatchedGroup = allPass([
-  pipe(nthArg(0), definesGroup),
-  pipe((binding, request) => definedGroup(binding) !== declaredGroup(request)),
-]);
-
-export const mismatchedVersion = allPass([
-  pipe(nthArg(0), definesVersion),
-  pipe((binding, request) => definedVersion(binding) !== declaredVersion(request)),
-]);
-
-export const mismatchedKind = allPass([
-  pipe(nthArg(0), definesKind),
-  pipe((binding, request) => definedKind(binding) !== declaredKind(request)),
 ]);
